Silence shellcheck warnings.
SC2039: local is not defined for sh but everything tested against implements it anyway.

SC2164: checks are done to be sure paths exist before cd is invoked.

SC2155: apparently local triggers this as well as export.

source following: since the output of the cache file is generated in the file before this it is already checked.