Add support for source C/C++ output
User who want to use autorevision for a project have to generate a
header file. However, this practice harmfull. Indeed, it is tempting
to include version.h in most other source files. Consequently, all
project need to be rebuild when version change.

Using a plain source file is a better practice. It allows to only
recompile one file on version change.

In order to avoid compiler warning, autorevision variables have to be
declared. We provide contribs/autorevision_declr.h for that.